+++ src/app/view/endpoints/clusters/cluster/organization/detail/users/organization-detail.users.module.js
@@ -33,14 +33,16 @@
     '$log',
     'app.utils.utilsService',
     'app.view.endpoints.clusters.cluster.manageUsers',
-    'app.view.endpoints.clusters.cluster.rolesService'
+    'app.view.endpoints.clusters.cluster.rolesService',
+    'app.event.eventService'
   ];
 
-  function OrganizationUsersController(modelManager, $stateParams, $state, $log, utils, manageUsers, rolesService) {
+  function OrganizationUsersController(modelManager, $stateParams, $state, $log, utils, manageUsers, rolesService, eventService) {
     var that = this;
 
     this.guid = $stateParams.guid;
     this.users = [];
+    this.removingSpace = {};
     this.usersModel = modelManager.retrieve('cloud-foundry.model.users');
 
     this.organizationModel = modelManager.retrieve('cloud-foundry.model.organization');
@@ -103,9 +105,7 @@
         name: gettext('Manage Roles'),
         disabled: true,
         execute: function (aUser) {
-          manageUsers.show(that.guid, [aUser], false).result.then(function () {
-            refreshUsers();
-          });
+          return manageUsers.show(that.guid, [aUser], false).result;
         }
       },
       {
@@ -133,16 +133,17 @@
 
     this.removeSpaceRole = function (user, spaceRole) {
       var space = spaceRole.space.details.space;
-      this.removingSpace = true;
+      var pillKey = space.entity.name + spaceRole.roleLabel;
+      if (this.removingSpace[pillKey]) {
+        return;
+      }
+      this.removingSpace[pillKey] = true;
       rolesService.removeSpaceRole(that.guid, space.entity.organization_guid, space.metadata.guid, user, spaceRole.role)
-        .then(function () {
-          return refreshUsers();
-        })
         .catch(function () {
           $log.error('Failed to remove role \'' + spaceRole.roleLabel + '\' for user \'' + user.entity.username + '\'');
         })
         .finally(function () {
-          that.removingSpace = false;
+          that.removingSpace[pillKey] = false;
         });
     };
 
@@ -164,6 +165,10 @@
     this.removeFromOrganization = function () {
 
     };
+
+    eventService.$on(eventService.events.ROLES_UPDATED, function () {
+      refreshUsers();
+    });
 
     // Ensure the parent state is fully initialised before we start our own init
     utils.chainStateResolve('endpoint.clusters.cluster.organization.detail.users', $state, init);
